backend/pyproject.toml
[project]
name = "shop-track"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "mitty1293", email = "k.fmichi@gmail.com" }
]
dependencies = [
    "django>=5.0.8",
    "djangorestframework>=3.15.2",
    "drf-spectacular>=0.27.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[tool.rye]
managed = true
virtual = true
dev-dependencies = []

[tool.rye.scripts]
devserver = { cmd = "python manage.py runserver", env-file = ".env" }
test = { cmd = "python manage.py test", env-file = ".env" }

[tool.ruff]
line-length = 119
indent-width = 4

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
extend-select = ["I", "N"]

[tool.ruff.lint.pydocstyle]
convention = "google"
